# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  errs.AppError:
    properties:
      code:
        description: |-
          Code 为唯一错误码，由 "DOMAIN.SNAKE_CASE" 格式的 ASCII 字符串构成。
          是前端做国际化映射（i18n key）和告警分组的依据。
        type: string
      details:
        description: |-
          Details 携带字段级校验错误明细，通常用于请求体（DTO）参数校验失败场景。
          如果无字段级错误，该 slice 应为 nil（JSON 序列化时被 omitempty 省略）。
        items:
          $ref: '#/definitions/errs.FieldDetail'
        type: array
      message:
        description: |-
          Message 是面向终端用户的中文提示文案，可直接渲染到 UI。
          要求：简洁、可理解、无敏感堆栈信息。前端根据此字段作为 fallback 文案展示。
        type: string
    type: object
  errs.FieldDetail:
    properties:
      field:
        description: |-
          Field 为请求参数名或 JSON path（如 "title"、"assignee.id"），
          前端可根据此字段将错误文案定位到对应表单控件。
        type: string
      reason:
        description: |-
          Reason 为该字段的具体错误原因（中文），如 "不能超过 255 个字符"、"格式不符合邮箱规范"。
          要求：简洁、无敏感信息、可直接渲染。
        type: string
    type: object
...
  webhook.Webhook:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      events:
        items:
          type: string
        type: array
      id:
        type: integer
      is_active:
        type: boolean
      last_error:
        type: string
      last_status:
        type: string
      last_triggered:
        type: string
      name:
        type: string
      project_id:
        type: integer
      target_url:
        type: string
      unhealthy_at:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 项目管理平台 Ydsz Plane 的 REST API（需求/任务/缺陷/迭代/版本/自动化/Webhook/知识库/仪表盘/全局搜索/通知等）。
  title: Ydsz Plane API
  version: 1.0.0
paths:
...
  /webhooks:
    get:
      description: 列出工作空间下的全部 Webhook 订阅，支持分页与项目过滤
      parameters:
      - description: 项目 ID
        in: query
        name: project_id
        type: integer
      - description: 每页数 (1-100)
        in: query
        name: limit
        type: integer
      - description: 偏移
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 列出 Webhook
      tags:
      - webhook
    post:
      consumes:
      - application/json
      description: 在工作空间下新建一个 Webhook 订阅，创建时返回 secret（后续不再返回）
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 创建 Webhook
      tags:
      - webhook
  /webhooks/{webhook_id}:
    delete:
      description: 删除指定 Webhook 订阅
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: 删除 Webhook
      tags:
      - webhook
    get:
      description: 返回单个 Webhook 的完整配置（不含 secret）
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.Webhook'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 获取 Webhook 详情
      tags:
      - webhook
    patch:
      consumes:
      - application/json
      description: PATCH 更新 Webhook 的 URL / 事件 / 名称 / 启用状态
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.Webhook'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 更新 Webhook
      tags:
      - webhook
  /webhooks/{webhook_id}/logs:
    get:
      description: 分页查询指定 Webhook 的投递日志，支持状态/事件类型过滤
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: integer
      - description: 投递状态
        in: query
        name: status
        type: string
      - description: 事件类型
        in: query
        name: event_type
        type: string
      - description: 每页数
        in: query
        name: limit
        type: integer
      - description: 偏移
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 列出投递日志
      tags:
      - webhook
  /webhooks/{webhook_id}/logs/{log_id}/retry:
    post:
      description: 手动重投指定投递日志（回查 domain_events 重建原始事件并同步重投）
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: integer
      - description: 日志 ID
        in: path
        name: log_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 重投日志
      tags:
      - webhook
  /webhooks/{webhook_id}/pause:
    post:
      description: '暂停 Webhook 投递（等价于 PATCH { is_active: false }）'
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.Webhook'
      summary: 暂停投递
      tags:
      - webhook
  /webhooks/{webhook_id}/resume:
    post:
      description: '恢复 Webhook 投递（等价于 PATCH { is_active: true }）'
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.Webhook'
      summary: 恢复投递
      tags:
      - webhook
  /webhooks/{webhook_id}/test:
    post:
      description: 向 Webhook 目标 URL 发送一条测试事件（ping）
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 测试投递
      tags:
      - webhook
...
securityDefinitions:
  Bearer:
    description: '"格式: Bearer {access_token}。登录/注册接口返回 access_token 后填入。"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"